+++ src/Application/VsAuth.cpp
@@ -34,12 +34,29 @@ VsAuth *VsAuth::instance()
 	return _instance;
 }
 
-bool VsAuth::isLoggedIn()
+bool VsAuth::hasValidToken()
 {
 	VsSettings *oSettings	= VsSettings::instance();
 	QVariant authExpireTime	= oSettings->value( SettingsKeys["AUTH_PAYLOAD"], SettingsGroups["authentication"] ).toHash().value( "tokenExpired" );
+	bool tokenNotExpired	= QDateTime::currentSecsSinceEpoch() <= authExpireTime.toInt();
 
-	return QDateTime::currentSecsSinceEpoch() <= authExpireTime.toInt();
+	return tokenNotExpired;
+}
+
+bool VsAuth::isLoggedIn()
+{
+	bool validToken	= hasValidToken();
+
+	if ( ! validToken ) {
+		VsSettings *oSettings	= VsSettings::instance();
+		QString refreshToken	= oSettings->value( SettingsKeys["REFRESH_TOKEN"], SettingsGroups["authentication"] ).toString();
+		if ( ! refreshToken.isEmpty() ) {
+			refreshAuthToken( refreshToken );
+			return true;
+		}
+	}
+
+	return validToken;
 }
 
 bool VsAuth::login( QString username, QString password )
@@ -89,4 +106,20 @@ void VsAuth::handleAuthResult( CommandState *state )
 	VsSettings::instance()->setValue( SettingsKeys["REFRESH_TOKEN"], refreshToken, SettingsGroups["authentication"] );
 
 	emit loginCheckFinished( state );
+}
+
+bool VsAuth::refreshAuthToken( QString refreshToken )
+{
+	QString strUrlLogin		= VsApplication::instance()->apiUrl().append( "/token/refresh" );
+	//qDebug() << strUrlLogin;
+
+	HttpRequestInput *input	= new HttpRequestInput( strUrlLogin, "POST" );
+	input->requestType	= REQUEST_TYPE_JSON;
+
+	input->addVar( "refresh_token", refreshToken );
+
+	//HttpRequestWorker::instance()->setObjectName( RequestTypes[LOGIN_CHECK] );
+	HttpRequestWorker::instance()->execute( input, HttpRequests["REFRESH_TOKEN_REQUEST"] );
+
+	return true;
 }