+++ src/Widget/Settings/AbstractSettingsWidget.cpp
@@ -1,0 +1,8 @@
+#include "AbstractSettingsWidget.h"
+
+#include "Application/VsSettings.h"
+
+AbstractSettingsWidget::AbstractSettingsWidget( QWidget *parent ) : QWidget( parent )
+{
+	settings	= VsSettings::instance();
+}