+++ src/Widget/SystemTrayMenu.cpp
@@ -106,13 +106,21 @@ void SystemTrayMenu::_createToolBar()
 
 	QMenu *profileMenu 	= new QMenu( "Profile" );
 
+	QAction *actionPreferences = new QAction( tr("&Preferences" ), this );
+	actionPreferences->setStatusTip( tr( "Open Preferences Dialog" ) );
+	actionPreferences->setIcon( QIcon( ":/Resources/icons/settings.svg" ) );
+	connect( actionPreferences, SIGNAL( triggered() ), this, SLOT( showSettingsWidget() ) );
+	profileMenu->addAction( actionPreferences );
+
 	QAction *logoutAct = new QAction( tr("&Sign Out" ), this );
 	logoutAct->setStatusTip( tr( "Sign Out From API" ) );
+	logoutAct->setIcon( QIcon( ":/Resources/icons/logout.svg" ) );
 	connect( logoutAct, SIGNAL( triggered() ), this, SLOT( logout() ) );
 	profileMenu->addAction( logoutAct );
 
 	QAction *quitAct = new QAction( tr("&Quit" ), this );
 	quitAct->setStatusTip( tr( "Quit Application" ) );
+	quitAct->setIcon( QIcon( ":/Resources/icons/close.svg" ) );
 	connect( quitAct, &QAction::triggered, QCoreApplication::instance(), &QCoreApplication::quit, Qt::QueuedConnection );
 	profileMenu->addAction( quitAct );
 
@@ -353,6 +361,14 @@ void SystemTrayMenu::openWebGuitarProFolder()
 	//qDebug() << WebGuitarProFolder;
 
 	QDesktopServices::openUrl( QUrl::fromLocalFile( WebGuitarProFolder ) );
+}
+
+void SystemTrayMenu::showSettingsWidget()
+{
+	wdgSettings = new SettingsWindow( this );
+	wdgSettings->setWindowFlags( Qt::Window );
+
+	wdgSettings->show();
 }
 
 void SystemTrayMenu::testFileUpload()